Move the test_utils.go file to its own package and renamed to shared.go
